Refactoring & Fixed CSS
