+++ src/Components/DisplayNote/DisplayNote.css
@@ -1,10 +1,11 @@
 .display-note {
-  margin: 0 0 2rem 2rem;
+  margin: 0 2rem 2rem 2rem;
   border-radius: 10px;
-  padding: 1rem;
+  padding: 2rem;
   min-height: 15rem;
   display: flex;
   flex-direction: column;
+  width: fit-content;
 }
 .note-input-container {
   align-items: center;
