+++ src/Components/TrashNote/TrashNote.css
@@ -1,4 +1,4 @@
-.display-note {
+.trash-display-note {
   margin: 0 0 2rem 2rem;
   border-radius: 10px;
   padding: 1rem;
