+++ src/Pages/Pages/ArchivePage/ArchivePage.css
@@ -19,12 +19,15 @@
 }
 .note-content-container {
   display: flex;
+  margin-top: 2rem;
 }
 .archive-note-container {
   display: block;
   width: 100%;
   height: 100%;
 }
-.main-page-title {
-  margin: 1rem 0 1rem 0;
+.archive-page-title {
+  margin: 1rem;
+  padding-left: 20rem;
+  white-space: nowrap;
 }
