+++ src/Pages/Pages/ArchivePage/ArchivePage.jsx
@@ -1,8 +1,6 @@
 import { Sidebar, ArchiveNote } from "../../../Components";
 import "./ArchivePage.css";
-import { useTrash } from "../../../context/trash-context";
-import { useNote } from "../../../context/note-context";
-import { useArchive } from "../../../context/archive-context";
+import { useTrash, useNote, useArchive } from "../../../context/context";
 export function ArchivePage() {
   const { trashNote, setTrashNote } = useTrash();
   const { notes, setNotes } = useNote();
@@ -13,7 +11,11 @@ export function ArchivePage() {
       <div className="note-content-container">
         <Sidebar />
         <div className="archive-note-container">
-          {archiveNote.length > 0 ? "" : <h3>Archived Notes!!!</h3>}
+          {archiveNote.length > 0 ? (
+            ""
+          ) : (
+            <h2 className="archive-page-title">No Archived Notes!!!</h2>
+          )}
           <div className="note-list">
             {archiveNote.map((item) => (
               <ArchiveNote key={item._id} {...item} />
