--- src/Pages/Pages/LabelPage/LabelPage.jsx
@@ -1,11 +1,0 @@
-import { Sidebar } from "../../../Components";
-import "./LabelPage.css";
-
-export function LabelPage() {
-  return (
-    <div>
-      <h1>Its a Label Page</h1>
-      <Sidebar />
-    </div>
-  );
-}
