+++ src/Pages/Pages/NotePage/NotePage.css
@@ -1,5 +1,5 @@
 .style-notepage {
-  max-width: 68.75rem;
+  max-width: 960px;
   margin-right: auto;
   padding-right: 15px;
   padding-left: 15px;
@@ -22,11 +22,14 @@
   display: flex;
 }
 .note-edit-container {
-  margin-left: 12rem;
+  margin-left: 5rem;
   margin-top: 2rem;
+  padding-left: 15rem;
 }
 .main-page-title {
-  margin: 1rem 0 1rem 0;
+  margin: 2rem;
+  padding-left: 50rem;
+  white-space: nowrap;
 }
 .sort-filter {
   display: flex;
