+++ src/Pages/Pages/TrashPage/TrashPage.css
@@ -19,6 +19,8 @@
   width: 100%;
   height: 100%;
 }
-.main-page-title {
-  margin: 1rem 0 1rem 0;
+.archive-page-title {
+  margin: 1rem;
+  padding-left: 30rem;
+  white-space: nowrap;
 }
