+++ src/Pages/Pages/TrashPage/TrashPage.jsx
@@ -1,6 +1,6 @@
 import { Sidebar, TrashNote } from "../../../Components";
 import "./TrashPage.css";
-import { useTrash } from "../../../context/trash-context";
+import { useTrash } from "../../../context/context";
 export function TrashPage() {
   const { trashNote, setTrashNote } = useTrash();
   console.log("Comes");
@@ -9,7 +9,11 @@ export function TrashPage() {
       <div className="note-content-container">
         <Sidebar />
         <div className="note-trash-container">
-          {trashNote.length > 0 ? "" : <h3>Trash is Empty!!!</h3>}
+          {trashNote.length > 0 ? (
+            ""
+          ) : (
+            <h2 className="trash-page-title">Trash Empty!!!</h2>
+          )}
           <div className="note-list">
             {trashNote.map((item) => (
               <TrashNote key={item._id} {...item} />
