+++ src/Pages/Pages/index.js
@@ -5,6 +5,5 @@ export { SignupPage } from "./SignupPage/SignupPage";
 export { HomePage } from "./HomePage/HomePage";
 export { LandingPage } from "./LandingPage/LandingPage";
 export { NotePage } from "./NotePage/NotePage";
-export { LabelPage } from "./LabelPage/LabelPage";
 export { ArchivePage } from "./ArchivePage/ArchivePage";
 export { TrashPage } from "./TrashPage/TrashPage";
