+++ src/context/context.js
@@ -1,0 +1,18 @@
+import { useAuth, AuthProvider } from "./auth-context";
+import { useNote, NoteProvider } from "./note-context";
+import { useFilter, FilterProvider } from "./filter-context";
+import { useTrash, TrashProvider } from "./trash-context";
+import { useArchive, ArchiveProvider } from "./archive-context";
+
+export {
+  useAuth,
+  AuthProvider,
+  useNote,
+  NoteProvider,
+  useFilter,
+  FilterProvider,
+  useArchive,
+  ArchiveProvider,
+  useTrash,
+  TrashProvider,
+};
